// apps/api/src/ai/vector-api.d.ts
export interface components {
    schemas: {
// ...
        GetDocumentsResponse: {
            /** @example success */
            status: string;
            collection: string;
            total?: number;
            limit?: number;
            offset?: number;
            data: components["schemas"]["DocumentsData"];
        };
        DocumentActionResponse: {
            /** @example success */
            status: string;
            /** @example upserted */
            action?: string;
            /** @example rops_innovations */
            collection: string;
            ids: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}